none-generic SetAuthSecretCreated function moved to authtoken package
